Correctly parse shorter completions from PCIe HIP
CycloneIV PCIe HIP can return data in the second cycle if the length is 1
DWORD and the address is divisible by 4, but not by 8 (i.e. bit 2 of the
lower address is set), because padding would have to be inserted into one
half of the bus in cycle 2, and into the other half in cycle 3.

This checks for length and EOP, as a shortcut.